feat: 회원 메세지 시스템 예시 작성
