+++ module/database/build.gradle
@@ -1,33 +1,20 @@
+// module/database/build.gradle
 plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.4.3'
+    id 'java-library'
     id 'io.spring.dependency-management' version '1.1.7'
 }
 
 group = 'hey.used.module.database'
-version = '0.0.1-SNAPSHOT'
+version = '0.0.1'
 
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-}
+java { toolchain { languageVersion = JavaLanguageVersion.of(21) } }
+repositories { mavenCentral() }
 
-configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
-}
-
-repositories {
-    mavenCentral()
+dependencyManagement {
+    imports { mavenBom "org.springframework.boot:spring-boot-dependencies:3.4.3" }
 }
 
 dependencies {
-
-    runtimeOnly 'com.mysql:mysql-connector-j:8.3.0'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
+    runtimeOnly 'com.mysql:mysql-connector-j:9.1.0'             // 또는 8.x
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
