--- module/kafka/producer/src/main/java/kafka/producer/GroupId.java
@@ -1,12 +1,0 @@
-package kafka.producer;
-
-public enum GroupId {
-
-
-    SIGN_UPED,
-
-
-
-
-
-}
