--- module/kafka/producer/src/main/java/kafka/producer/Topic.java
@@ -1,13 +1,0 @@
-package kafka.producer;
-
-public enum Topic {
-
-    // 도메인
-
-    USER,
-
-
-
-
-
-}
