--- user/api/src/main/java/hey/car/user/CarUserApplication.java
@@ -1,13 +1,0 @@
-package hey.car.user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CarUserApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CarUserApplication.class, args);
-    }
-
-}
