+++ user/api/src/main/java/hey/car/user/api/common/audit/BaseAuditEntity.java
@@ -1,0 +1,37 @@
+package hey.car.user.api.common.audit;
+
+import java.time.Instant;
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseAuditEntity {
+
+    @CreatedDate
+    @Column(nullable = false, updatable = false)
+    protected Instant createdAt;
+
+    @LastModifiedDate
+    @Column(nullable = false)
+    protected Instant updatedAt;
+
+    @CreatedBy
+    @Column(updatable = false, length = 100)
+    protected String createdBy;
+
+    @LastModifiedBy
+    @Column(length = 100)
+    protected String updatedBy;
+
+    public Instant getCreatedAt() { return createdAt; }
+    public Instant getUpdatedAt() { return updatedAt; }
+    public String getCreatedBy() { return createdBy; }
+    public String getUpdatedBy() { return updatedBy; }
+}
