+++ user/api/src/main/java/hey/car/user/api/common/audit/JpaAuditingConfig.java
@@ -1,0 +1,11 @@
+package hey.car.user.api.common.audit;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+
+
+}
