+++ user/api/src/main/java/hey/car/user/api/common/validation/Validator.java
@@ -1,0 +1,6 @@
+package hey.car.user.api.common.validation;
+
+public interface Validator {
+
+
+}
