+++ user/api/src/main/java/hey/car/user/api/user/entity/MessageTopic.java
@@ -1,0 +1,18 @@
+package hey.car.user.api.user.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum MessageTopic {
+
+    // 도메인
+
+    USER_CREATED("user.created"),;
+
+
+    private final String topic;
+
+
+}
