+++ user/api/src/main/java/hey/car/user/api/user/entity/Password.java
@@ -1,0 +1,34 @@
+package hey.car.user.api.user.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.Instant;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor(access = AccessLevel.PUBLIC)
+@AllArgsConstructor(access = AccessLevel.PUBLIC)
+public class Password {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true, length = 191)
+    private String username;
+
+    @Column(nullable = false, unique = true, length = 191)
+    private String password;
+
+    private Instant createdAt;
+
+    private Instant updatedAt;
+
+
+
+
+
+
+}
