+++ user/api/src/main/java/hey/car/user/api/user/entity/UserType.java
@@ -1,0 +1,14 @@
+package hey.car.user.api.user.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum UserType {
+
+    CUSTOMER,
+    SELLER
+
+    ;
+
+
+}
