+++ user/api/src/main/java/hey/car/user/api/user/repository/UserJpaRepository.java
@@ -1,0 +1,9 @@
+package hey.car.user.api.user.repository;
+
+import hey.car.user.api.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserJpaRepository  extends JpaRepository<User, Long> {
+
+    boolean existsByUsername(String username);
+}
