--- user/api/src/main/java/hey/car/user/common/audit/JpaAuditingConfig.java
@@ -1,11 +1,0 @@
-package hey.car.user.common.audit;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-
-
-}
