--- user/api/src/main/java/hey/car/user/common/validation/UserValidator.java
@@ -1,17 +1,0 @@
-package hey.car.user.common.validation;
-
-import hey.car.user.user.repository.UserJpaRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class UserValidator {
-
-    private final UserJpaRepository userRepository;
-
-    public boolean exist(String username) {
-        return userRepository.existsByUsername(username);
-    }
-
-}
