--- user/api/src/main/java/hey/car/user/common/validation/Validator.java
@@ -1,6 +1,0 @@
-package hey.car.user.common.validation;
-
-public interface Validator {
-
-
-}
