--- user/api/src/main/java/hey/car/user/user/controller/dto/UserCreateReq.java
@@ -1,13 +1,0 @@
-package hey.car.user.user.controller.dto;
-
-
-import hey.car.user.user.entity.UserType;
-
-public record UserCreateReq(
-        String username,
-        String password,
-        String email,
-        String phone,
-        UserType type
-)
-{ }
