--- user/api/src/main/java/hey/car/user/user/entity/UserType.java
@@ -1,14 +1,0 @@
-package hey.car.user.user.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum UserType {
-
-    CUSTOMER,
-    SELLER
-
-    ;
-
-
-}
