--- user/api/src/main/java/hey/car/user/user/repository/PasswordJpaRepository.java
@@ -1,8 +1,0 @@
-package hey.car.user.user.repository;
-
-import hey.car.user.user.entity.Password;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PasswordJpaRepository extends JpaRepository<Password, Long> {
-
-}
