+++ user/message-worker/src/main/resources/application.yml
@@ -1,0 +1,16 @@
+
+server:
+  port: 10502
+
+
+slack:
+  webhook-url: https://hooks.slack.com/services/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
+
+spring:
+  jpa:
+    hibernate:
+      ddl-auto: create
+    show-sql: true
+    properties:
+      hibernate.format_sql: false
+      hibernate.jdbc.time_zone: Asia/Seoul
